Add basic support for optipng.
optipng serves as an alternative to pngout, which is (unfortunately) not
Open Source/Free Software and may not be available on some systems
(especially those that do not run on x86/amd64 platforms, like ARM and its
variants or PowerPC).

Coupled with a second run by advancecomp (advpng/advdef), optipng is very
frequently competitive with pngout and even pngout benefits from advpng
and/or advdef in my experience.

This commit puts optipng disabled by default and users have to explicitly
enable it via `--use-optipng=yes`.

This default may be revised in a future version of pdfsizeopt and be enabled
by default, especially if followed by advpng/advdef, which give substantial
gains if the user is willing to incur in a speed penalty for the exchange of
a better compression of images.

Signed-off-by: Rogério Theodoro de Brito <rbrito@ime.usp.br>